requirements.txt
albumentations==1.1.0
appdirs==1.4.3
apturl==0.5.2
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.2
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.7
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
httplib2==0.14.0
idna==2.8
image==1.5.33
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==1.5.0
iotop==0.6
joblib==1.0.1
keyring==18.0.1
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
liac-arff==2.5.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.4.3
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
networkx==2.6.3
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.3.0
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
PyWavelets==1.1.1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
qudida==0.0.4
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
scikit-image==0.18.3
scikit-learn==1.0
scipy==1.7.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Shapely==1.7.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
systemd-python==234
threadpoolctl==3.0.0
tifffile==2021.8.30
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.7.2
wadllib==1.3.3
xkit==0.0.0
youtube-dl==2020.3.24
zipp==1.0.0
